[#18] separating websocke and api directories
