+++ config/websocket/gunicorn.conf.toml
@@ -1,0 +1,5 @@
+bind = "0.0.0.0:8001"
+workers = 1
+worker_class = "uvicorn.workers.UvicornWorker"
+accesslog = "-"
+errorlog = "-"
