+++ mas/api/auth/exception/auth_exception.py
