+++ mas/api/mas_api_application.py
@@ -1,0 +1,11 @@
+import os
+
+from mas.api.api_initializer import APIInitializer
+from mas.utils.controller_utils import create_app
+from mas.utils.gunicorn_utils import MASMultiProcessingServer
+
+APIInitializer(os.getenv("PHASE", "local"))
+
+app = create_app("./mas/api")
+
+MASMultiProcessingServer(app, "./config/api/gunicorn.conf.toml").run()
