--- mas/mas_application.py
@@ -1,6 +1,0 @@
-from mas.utils.controller_utils import create_app
-from mas.utils.gunicorn_utils import MASMultiThreadingServer
-
-app = create_app()
-
-MASMultiThreadingServer(app).run()
