+++ mas/websocket/script/dto/script_response.py
